Add lower instruction noop

+++ neko_tests/vpu/opcode_tests/vpu_upper_instruction_utils.hpp
@@ -1,7 +1,7 @@
 #ifndef VPU_UPPER_INSTRUCTION_UTILS_HPP
 #define VPU_UPPER_INSTRUCTION_UTILS_HPP
 
-void addInstructionToVector(vector<uint8_t> * instructions, uint32_t bitFlags, uint32_t destFlags, uint8_t ftRegID, uint8_t fsRegID, uint8_t fdRegID, uint16_t opCode);
+void addInstructionToVector(vector<uint8_t> * instructions, uint32_t bitFlags, uint32_t destFlags, uint8_t ftRegID, uint8_t fsRegID, uint8_t fdRegID, uint32_t opCode);
 void addNOPHalfInstructionToVector(vector<uint8_t> * instructions);
 void addNOPFullInstructionToVector(vector<uint8_t> * instructions);
 void addSingleUpperInstruction(vector<uint8_t> * instructions, uint32_t bitFlags, uint32_t destFlags, uint8_t ftRegID, uint8_t fsRegID, uint8_t fdRegID, uint16_t opCode);
